iOS versions of VVBufferPool and VVISFKit
most of the functionality of VVBufferPool is available in iOS, though certain texture formats/types may not be available and some functionality may be missing.  VVISFKit appears to be fully functional in our tests, though it requires GL ES 3 for some of the more complex shaders to work as expected.  all of this requires iOS 8 or later- also deleted all the old static libs for iOS and replaced them with frameworks.