Moved some src files around
This was done to improve compatibility with other build systems that may not support xcode’s headermap.  The fundamental issue is that #import paths need to be consistent between both the source tree and the deployed target- if they aren’t (as was the case prior to this commit), external apps that attempt to #import headers will be looking for them in the wrong place!